# PERF: Optimize loadtxt usecols.